create useRequest hook
